Rudimentary GameWindow
Not entirely sure how we maintain eclipse project folders here as
compared to VS...so I just dropped the whole project folder in there.

As of yet, it simply displays the look of a JFrame containing a JPanel
for a basic look of the game board and location of pieces at start.

Buttons for MainMenu and Pause are visible but do nothing since I'm not
sure what the other code looks like to link things properly and such.
Similarly, the board/pieces are nothing more than a visualization in
absolute coordinates within the frame/panel for now. They shall be
converted into more distinctive objects with useful properties (like in
the backlog) once I can see how our framework is intended to interact.