Some Board/Piece reconsolidation
-Piece was given its own class and appropriate method, plus a few
attributes that we will probably need to change.
-highlight() method was added to Piece. Currently, running the program
will just highlight all the pieces for demonstration purposes...since I
have NOT added a mouselistener yet. I'm not really sure what objects in
use are tangible for that purpose so I'll look at it on thursday maybe.

-Board initialization code was cleaned up a little by grouping the draws
into methods
-Board objects now have a 5x9 matrix array of Piece objects that should
be useful for actually setting up logistics of the game. [2][4] of the
matrix is not represented by a piece since that's the initial empty
spot. We need to decide how we want to approach that, use of a Piece
object with an "empty" field might be an easy way to go. That would
simplify drawing as well since it wouldn't necessarily need to be
conditional.
-Added a few placeholder methods
-Removed the unused imports...I recall we sorta debated whether to leave
them in, but they were put there by WindowBuilder most likely...and
WindowBuilder will just add again if necessary.
-