Restructured Board & Piece
Changes made to Board.java and Piece.java to make them more conducive to
invocation of paint() and repaint(). As such, board.repaint() will
redraw the board JPanel with any changes that were made, such as
visibility or highlighting of pieces. A test to demonstrate the update
ability was performed in the Pause Button ActionEvent(?) handler. Search
"//TESTING" to find this code snippet.

In regards to the clickable pieces aspect, I still haven't really
figured out what approach to use.