Sacrifice implemented and a few other things
Sacrifice is now a working part of the game. There's a toggle button to
use it in-game. A black circle with an X through it is the graphical
representation.

Changed the gameplay relevent files' main comment sections to
documentation comments to differentiate from notes to selves and code
commented out.

Started to add the options to host as P1 or P2 and as Human or
CPU...We'll finish that when John can look at it so I don't have to look
up the classes when he knows them already.